backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
PyGithub==2.3.0
google-generativeai==0.7.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
httpx==0.27.0
itsdangerous==2.2.0
python-multipart==0.0.9
llama-index-core==0.10.43
llama-index-llms-gemini==0.1.11
llama-index-embeddings-gemini==0.1.8
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
anyio==4.4.0
